refactor(bundle-source): Extract fs module
